echo-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.5.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
loguru==0.7.2
slowapi==0.1.9

# Testing
pytest==8.2.1
httpx==0.27.0
pytest-asyncio==0.23.7

